refactor(neon_framework): introduce storage library to handle all the storage need for Neon
