+++ packages/neon_framework/lib/src/settings/models/option.dart
@@ -7,6 +7,7 @@ import 'package:neon_framework/src/models/disposable.dart';
 import 'package:neon_framework/src/models/label_builder.dart';
 import 'package:neon_framework/src/settings/models/options_category.dart';
 import 'package:neon_framework/src/settings/models/storage.dart';
+import 'package:neon_framework/storage.dart';
 import 'package:rxdart/rxdart.dart';
 
 /// Listenable option that is persisted in the [SettingsStorage].
