+++ packages/neon_framework/lib/storage.dart
@@ -1,0 +1,8 @@
+/// Storage interfaces for the Neon app framework.
+///
+/// The `NeonStorage` manages all storage backends.
+library;
+
+export 'package:neon_framework/src/storage/app_storage.dart';
+export 'package:neon_framework/src/storage/single_value_store.dart';
+export 'package:neon_framework/src/storage/storage_manager.dart';
