+++ packages/neon_framework/test/storage_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:neon_framework/src/settings/models/storage.dart';
+import 'package:neon_framework/storage.dart';
 import 'package:neon_framework/testing.dart';
 import 'package:shared_preferences/shared_preferences.dart';
